requirements.txt
crewai
langchain_google_genai #to interact with any gemini models
load_dotenv #Any environment i'll be creating ill be able to use this
